+++ external_tests/test_attacks_on_cleverhans.py
@@ -64,7 +64,7 @@ np.random.seed(seed)
 random.seed(seed)
 tf.set_random_seed(seed)
 inputs = np.random.uniform(0, 1, size=(BATCH_SIZE, DIM_INPUT))
-targets = np.random.randint(0, NUM_CLASS, size=BATCH_SIZE)
+targets = np.random.randint(0, NUM_CLASS, size=BATCH_SIZE, dtype=np.int64)
 
 
 targets_onehot = np.zeros((BATCH_SIZE, NUM_CLASS), dtype='int')
